refactor(runtime): run Index and Git turns behind one Vault work executor (#197)
Reading one Index turn or one Git turn meant following seven to nine files:
the dispatch loop and readiness policy were inlined in the composition root,
and the turn bodies lived in the collection-runtime module. Both now live in
`src/vault_executor.rs`. `server.rs` keeps only the loop that takes the next
coordinator position and hands it to `VaultWorkExecutor::run`, then applies
`publish_outcome`.

Folds in #128: the three Git-capable source kinds each repeated the same
shell — credential resolution, mutation-lock acquisition, `spawn_blocking`,
panic mapping, outcome publication. `plan_git_turn` now supplies only what
differs, as a `GitTurnPlan`; the shell runs once. Its `GitTurnWork` makes the
checkout lease part of the variant, so a managed-Git turn cannot be built or
run without one. Acquisition order, lock scope, panic codes, retryability,
and published outcomes are unchanged for every source kind.

Folds in #127 in the form the maintainer chose: rather than merging the work
coordinator's per-Vault map into the Git scheduler's — the coordinator's
entries are dropped the moment a Vault goes idle, while the scheduler's must
survive a Vault's whole active lifetime to hold its checkout lease (#95) —
the overlap itself is removed. `ManagedGitScheduler::tick` no longer reads
coordinator state and then acts on it; `VaultWorkCoordinator::request_if_idle`
decides and enqueues under the one lock that owns the answer. `has_work` is
now a `#[cfg(test)]` observation with no production caller.

Interface change: VaultWorkCoordinator automatic-request contract
Producer boundary: Background work coordination (`src/vault_work.rs`)
Kind: additive plus breaking (one removal)
Old contract: `pub fn has_work(vault_id, kind) -> bool`, a read-only query a
  producer read and then acted on.
New contract: `pub fn request_if_idle(vault_id, kind) -> ScheduleResult`,
  which performs that check and the enqueue under one lock and never adds the
  guaranteed rerun `request` gives an already-active turn. `has_work` becomes
  `#[cfg(test)]`.
Consumer boundaries: Git synchronization (`git::ManagedGitScheduler::tick`) —
  the only production consumer, migrated. Vault work execution, runtime
  composition, and Vault collection management use it only from tests. No
  external consumers: the crate is not published and the type is not
  serialized, routed, or exposed over HTTP or MCP.
Compatibility/migration: none needed; every in-repository consumer is updated
  in this change. No observable HTTP, MCP, search, or Git behaviour changes —
  `tick`'s decision is identical, only now atomic.
Rollout/rollback: N/A — one binary, no persisted format, no staged rollout.
Evidence: from the repository root — `cargo fmt --all -- --check`,
  `cargo clippy --all-targets --all-features -- -D warnings`,
  `cargo test --all` (883 passed), `cargo test --all --features eval`
  (883 + 8 + 5 + 1 passed), `node scripts/check-module-map.mjs` (199 files,
  one assignment each), `just docs-freshness` + `just docs-freshness-ack`
  (13 notes reviewed against 8 changed surfaces). Live acceptance on
  `just dev-start messy`.
ADR check: ADR-13 (no trait, framework, or code generation added — a plain
  module), ADR-18 (no second execution lane; a Git turn still holds its
  Vault's mutation lock and never force-checks out), ADR-19 (orchestration
  stays in the core). Others N/A: no write path, cache model, auth, or
  retrieval change.

Closes #197 (#127 and #128 were already closed as absorbed into it).
