+++ docs/adr/README.md
@@ -214,7 +214,7 @@ Copy this for a new record:
 - **Context:** ADR-10 fixed optional Git sync as one debounced background task over the process-wide `VAULT_PATH`, started from the settings page and locking phase by phase against an instance-wide write lock. The Vault collection runtime (v2.5.0) changed the unit of work: each Vault carries its own definition (source kind, Git mode, credentials, poll interval, commit identity) and gets its own Git turn, requested through the shared `VaultWorkCoordinator` by the managed-Git scheduler, a manual sync or retry, or activation, and executed by runtime composition under that Vault's mutation lock. The old task still compiled and was still reachable from the settings Git lifecycle, which gates on a value the collection runtime never sets, so in production it answered 503; the architecture review of 2026-08-28 (#162) found the two mechanisms had diverged on locking granularity, drift handling, and recovery, and that the instance-wide Versioning console and `/api/git-status` reported only the dead task.
 - **Decision:** The per-Vault Git turn is the only Git synchronisation mechanism. One turn per Vault, one operation at a time, scheduled through the coordinator with no second execution lane; a turn holds that Vault's mutation lock for its whole blocking duration; the Vault's source kind and Git mode select the operation (acquire-or-reuse then synchronise a managed checkout; synchronise an existing checkout with its one configured remote; commit local history without touching a remote). ADR-10's safety decisions are restated for the turn: sync is optional and off per Vault by default; a turn never force-checks out over uncommitted manual edits, reporting a dirty working copy or a conflict as a structured, non-retryable error with the affected paths; local-history mode never contacts a remote; writes never block on sync except for the one Vault whose turn is running. The instance-wide debounced task, its handle, its write-record hook, and the instance-wide write lock are removed (#185). The boot-time parse of `HATCHDOOR_GIT_*` is deliberately kept, now serving only two purposes: refusing to start on a half-configured Git mode, and feeding the demo-mode posture refusal. No runtime behaviour reads its result. `HATCHDOOR_GIT_AUTHOR_NAME` and `HATCHDOOR_GIT_AUTHOR_EMAIL` remain as the commit-identity fallback for a Vault without its own and are read per turn (#181); the other `HATCHDOOR_GIT_*` keys remain only as first-boot import inputs until #82 closes.
 - **Consequences:** One Git code path to audit, and Git behaviour that is configured, reported, and controlled per Vault (status, error detail, `Sync now`, `Try again`). The per-turn lock is coarser than the old per-phase lock: a foreground write to a Vault waits while that Vault's turn runs, a trade accepted in #96 over a finer discipline. The commit-on-write debounce has no successor: the watcher's fixed debounce coalesces writes into the next turn, and the per-Vault poll interval answers how often to look at the remote. The settings page loses its instance-wide Versioning console and `/api/git-status` is retired (#183).
-- **Evidence:** `git/managed_task.rs` (`run_managed_git_turn`, `run_existing_git_remote_turn`, `ManagedGitScheduler`), `git/managed_sync.rs` (`synchronize_managed_checkout`), `git/managed_checkout.rs` (`acquire_or_reuse`), `git/sync.rs` (`run_local_history_git_turn`), `vault_runtime.rs` (`dispatch_managed_git_turn_with`, `publish_managed_git_turn_outcome`), `vault_registry.rs` (`VaultGitMode`, `poll_interval_secs`, `commit_identity`); #162, #181, #183; issues #94–#97, #132 for the per-Vault turn's history. The removal itself landed in #185: `git/task.rs` and `git/status.rs` are deleted outright, `git/sync.rs` keeps only the local commit path (`validate_repo`, `validate_local_repo`, `init_local_repo`, `commit_local`, `run_local_history_git_turn`) with its fetch/integrate/push, unpushed accounting, and merge-marker recovery gone, `AppState` loses `git_sync`, `vault_write_lock`, and `startup_git_config`, `server.rs` loses the sync-task shutdown drain but keeps the boot-time `HATCHDOOR_GIT_*` parse for configuration validation and the demo posture check (`server.rs:1010-1031`), and `handlers/settings.rs` loses `patch_settings_with_git_lifecycle` and `production_sync_ops`.
+- **Evidence:** `git/managed_task.rs` (`run_managed_git_turn`, `run_existing_git_remote_turn`, `ManagedGitScheduler`), `git/managed_sync.rs` (`synchronize_managed_checkout`), `git/managed_checkout.rs` (`acquire_or_reuse`), `git/sync.rs` (`run_local_history_git_turn`), `vault_executor.rs` (`dispatch_git_turn_with`, `plan_git_turn`, `publish_managed_git_turn_outcome`; extracted from `vault_runtime.rs` in #197), `vault_registry.rs` (`VaultGitMode`, `poll_interval_secs`, `commit_identity`); #162, #181, #183; issues #94–#97, #132 for the per-Vault turn's history. The removal itself landed in #185: `git/task.rs` and `git/status.rs` are deleted outright, `git/sync.rs` keeps only the local commit path (`validate_repo`, `validate_local_repo`, `init_local_repo`, `commit_local`, `run_local_history_git_turn`) with its fetch/integrate/push, unpushed accounting, and merge-marker recovery gone, `AppState` loses `git_sync`, `vault_write_lock`, and `startup_git_config`, `server.rs` loses the sync-task shutdown drain but keeps the boot-time `HATCHDOOR_GIT_*` parse for configuration validation and the demo posture check (`server.rs:1010-1031`), and `handlers/settings.rs` loses `patch_settings_with_git_lifecycle` and `production_sync_ops`.
 
 ## ADR-19 — Vault-qualified cores are the only seam an adapter crosses
